Move UserList to SeenUserList directory
